backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
pg8000>=1.31.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
python-multipart>=0.0.9
pydantic[email]>=2.9.0
python-dotenv>=1.0.1
celery>=5.4.0
redis>=5.0.0
reportlab>=4.2.0